+++ Forge1.12.2/src/main/java/com/Fishmod/mod_LavaCow/entities/EntityZombieMushroom.java
@@ -291,7 +291,7 @@ public class EntityZombieMushroom extends EntityZombie implements IAggressive{
     
     public int getSkin()
     {
-        return ((Integer)this.dataManager.get(SKIN_TYPE)).intValue();
+        return this.dataManager.get(SKIN_TYPE).intValue();
     }
 
     public void setSkin(int skinType)
