+++ Forge1.12.2/src/main/java/com/Fishmod/mod_LavaCow/entities/tameable/EntityScarecrow.java
@@ -307,7 +307,7 @@ public class EntityScarecrow  extends EntityFishTameable{
     
     public int getSkin()
     {
-        return ((Integer)this.dataManager.get(SKIN_TYPE)).intValue();
+        return this.dataManager.get(SKIN_TYPE).intValue();
     }
 
     public void setSkin(int skinType)
